+++ test/vitest/wait-for-coordinates.spec.ts
@@ -6,14 +6,13 @@ import {
   createUser,
   waitForCoordinates,
 } from "fake-geolocation";
-import { createPermissionStore, createPermissions } from "fake-permissions";
+import { createPermissionStore } from "fake-permissions";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { coordsA, coordsB, coordsC } from "../fixture/coords.js";
 
 describe("waitForCoordinates()", () => {
   const startTime = 100;
   let locationServices: MutableLocationServices;
-  let permissions: Permissions;
   let user: User;
   let geolocation: Geolocation;
 
@@ -27,18 +26,13 @@ describe("waitForCoordinates()", () => {
       initialStates: new Map([[{ name: "geolocation" }, "granted"]]),
     });
 
-    permissions = createPermissions({ permissionStore });
-
     user = createUser({ locationServices, permissionStore });
     user.jumpToCoordinates(coordsA);
 
     geolocation = createGeolocation({
       locationServices,
-      permissions,
-
-      async requestPermission(descriptor) {
-        return user.requestPermission(descriptor);
-      },
+      permissionStore,
+      user,
     });
   });